Split provider code into separate modules
This commit splits the code for the 3 existing providers into separate
modules: one for key management (containing the code for key
destruction, creation, import and export) and one for asymetric signing
(currently containing code for signing and verifying hashes).
