+++ src/providers/mbed_provider/mod.rs
@@ -2,13 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 use super::Provide;
 use crate::authenticators::ApplicationName;
-use crate::key_info_managers;
-use crate::key_info_managers::{KeyInfo, KeyTriple, ManageKeyInfo};
+use crate::key_info_managers::{KeyTriple, ManageKeyInfo};
 use constants::PSA_SUCCESS;
 use derivative::Derivative;
-use log::{error, info, warn};
+use log::error;
 use parsec_interface::operations::list_providers::ProviderInfo;
-use parsec_interface::operations::psa_key_attributes::KeyAttributes;
 use parsec_interface::operations::{
     list_opcodes, psa_destroy_key, psa_export_public_key, psa_generate_key, psa_import_key,
     psa_sign_hash, psa_verify_hash,
@@ -34,8 +32,10 @@ mod psa_crypto_binding {
     include!(concat!(env!("OUT_DIR"), "/psa_crypto_bindings.rs"));
 }
 
+mod asym_sign;
 #[allow(dead_code)]
 mod constants;
+mod key_management;
 mod utils;
 
 type LocalIdStore = HashSet<psa_key_id_t>;
@@ -71,78 +71,6 @@ pub struct MbedProvider {
     // assigned at any time.
     #[derivative(Debug = "ignore")]
     key_slot_semaphore: Semaphore,
-}
-
-/// Gets a PSA Key ID from the Key Info Manager.
-/// Wrapper around the get method of the Key Info Manager to convert the key ID to the psa_key_id_t
-/// type.
-fn get_key_id(key_triple: &KeyTriple, store_handle: &dyn ManageKeyInfo) -> Result<psa_key_id_t> {
-    match store_handle.get(key_triple) {
-        Ok(Some(key_info)) => {
-            if key_info.id.len() == 4 {
-                let mut dst = [0; 4];
-                dst.copy_from_slice(&key_info.id);
-                Ok(u32::from_ne_bytes(dst))
-            } else {
-                error!("Stored Key ID is not valid.");
-                Err(ResponseStatus::KeyInfoManagerError)
-            }
-        }
-        Ok(None) => Err(ResponseStatus::PsaErrorDoesNotExist),
-        Err(string) => Err(key_info_managers::to_response_status(string)),
-    }
-}
-
-/// Creates a new PSA Key ID and stores it in the Key Info Manager.
-fn create_key_id(
-    key_triple: KeyTriple,
-    key_attributes: KeyAttributes,
-    store_handle: &mut dyn ManageKeyInfo,
-    local_ids_handle: &mut LocalIdStore,
-) -> Result<psa_key_id_t> {
-    let mut key_id = rand::random::<psa_key_id_t>();
-    while local_ids_handle.contains(&key_id)
-        || key_id == 0
-        || key_id > constants::PSA_MAX_PERSISTENT_KEY_IDENTIFIER
-    {
-        key_id = rand::random::<psa_key_id_t>();
-    }
-    let key_info = KeyInfo {
-        id: key_id.to_ne_bytes().to_vec(),
-        attributes: key_attributes,
-    };
-    match store_handle.insert(key_triple.clone(), key_info) {
-        Ok(insert_option) => {
-            if insert_option.is_some() {
-                warn!("Overwriting Key triple mapping ({})", key_triple);
-            }
-            let _ = local_ids_handle.insert(key_id);
-
-            Ok(key_id)
-        }
-        Err(string) => Err(key_info_managers::to_response_status(string)),
-    }
-}
-
-fn remove_key_id(
-    key_triple: &KeyTriple,
-    key_id: psa_key_id_t,
-    store_handle: &mut dyn ManageKeyInfo,
-    local_ids_handle: &mut LocalIdStore,
-) -> Result<()> {
-    match store_handle.remove(key_triple) {
-        Ok(_) => {
-            let _ = local_ids_handle.remove(&key_id);
-            Ok(())
-        }
-        Err(string) => Err(key_info_managers::to_response_status(string)),
-    }
-}
-
-fn key_info_exists(key_triple: &KeyTriple, store_handle: &dyn ManageKeyInfo) -> Result<bool> {
-    store_handle
-        .exists(key_triple)
-        .or_else(|e| Err(key_info_managers::to_response_status(e)))
 }
 
 impl MbedProvider {
@@ -181,7 +109,7 @@ impl MbedProvider {
             match store_handle.get_all(ProviderID::MbedCrypto) {
                 Ok(key_triples) => {
                     for key_triple in key_triples.iter().cloned() {
-                        let key_id = match get_key_id(key_triple, &*store_handle) {
+                        let key_id = match key_management::get_key_id(key_triple, &*store_handle) {
                             Ok(key_id) => key_id,
                             Err(response_status) => {
                                 error!("Error getting the Key ID for triple:\n{}\n(error: {}), continuing...", key_triple, response_status);
@@ -249,63 +177,7 @@ impl Provide for MbedProvider {
         app_name: ApplicationName,
         op: psa_generate_key::Operation,
     ) -> Result<psa_generate_key::Result> {
-        info!("Mbed Provider - Create Key");
-        let _semaphore_guard = self.key_slot_semaphore.access();
-        let key_name = op.key_name;
-        let key_attributes = op.attributes;
-        let key_triple = KeyTriple::new(app_name, ProviderID::MbedCrypto, key_name);
-        let mut store_handle = self
-            .key_info_store
-            .write()
-            .expect("Key store lock poisoned");
-        let mut local_ids_handle = self.local_ids.write().expect("Local ID lock poisoned");
-        if key_info_exists(&key_triple, &*store_handle)? {
-            return Err(ResponseStatus::PsaErrorAlreadyExists);
-        }
-        let key_id = create_key_id(
-            key_triple.clone(),
-            key_attributes,
-            &mut *store_handle,
-            &mut local_ids_handle,
-        )?;
-
-        let key_attrs = utils::convert_key_attributes(&key_attributes, key_id).or_else(|e| {
-            remove_key_id(
-                &key_triple,
-                key_id,
-                &mut *store_handle,
-                &mut local_ids_handle,
-            )?;
-            error!("Failed converting key attributes.");
-            Err(e)
-        })?;
-
-        let _guard = self
-            .key_handle_mutex
-            .lock()
-            .expect("Grabbing key handle mutex failed");
-
-        // Safety:
-        //   * at this point the provider has been instantiated so Mbed Crypto has been initialized
-        //   * self.key_handle_mutex prevents concurrent accesses
-        //   * self.key_slot_semaphore prevents overflowing key slots
-        let mut key_handle = unsafe { KeyHandle::generate(&key_attrs) }.or_else(|e| {
-            remove_key_id(
-                &key_triple,
-                key_id,
-                &mut *store_handle,
-                &mut local_ids_handle,
-            )?;
-            error!("Generate key status: {}", e);
-            Err(e)
-        })?;
-
-        // Safety: same conditions than above.
-        unsafe {
-            key_handle.close()?;
-        }
-
-        Ok(psa_generate_key::Result {})
+        self.psa_generate_key_internal(app_name, op)
     }
 
     fn psa_import_key(
@@ -313,64 +185,7 @@ impl Provide for MbedProvider {
         app_name: ApplicationName,
         op: psa_import_key::Operation,
     ) -> Result<psa_import_key::Result> {
-        info!("Mbed Provider - Import Key");
-        let _semaphore_guard = self.key_slot_semaphore.access();
-        let key_name = op.key_name;
-        let key_attributes = op.attributes;
-        let key_data = op.data;
-        let key_triple = KeyTriple::new(app_name, ProviderID::MbedCrypto, key_name);
-        let mut store_handle = self
-            .key_info_store
-            .write()
-            .expect("Key store lock poisoned");
-        let mut local_ids_handle = self.local_ids.write().expect("Local ID lock poisoned");
-        if key_info_exists(&key_triple, &*store_handle)? {
-            return Err(ResponseStatus::PsaErrorAlreadyExists);
-        }
-        let key_id = create_key_id(
-            key_triple.clone(),
-            key_attributes,
-            &mut *store_handle,
-            &mut local_ids_handle,
-        )?;
-
-        let key_attrs = utils::convert_key_attributes(&key_attributes, key_id).or_else(|e| {
-            remove_key_id(
-                &key_triple,
-                key_id,
-                &mut *store_handle,
-                &mut local_ids_handle,
-            )?;
-            error!("Failed converting key attributes.");
-            Err(e)
-        })?;
-
-        let _guard = self
-            .key_handle_mutex
-            .lock()
-            .expect("Grabbing key handle mutex failed");
-
-        // Safety:
-        //   * at this point the provider has been instantiated so Mbed Crypto has been initialized
-        //   * self.key_handle_mutex prevents concurrent accesses
-        //   * self.key_slot_semaphore prevents overflowing key slots
-        let mut key_handle = unsafe { KeyHandle::import(&key_attrs, key_data) }.or_else(|e| {
-            remove_key_id(
-                &key_triple,
-                key_id,
-                &mut *store_handle,
-                &mut local_ids_handle,
-            )?;
-            error!("Import key status: {}", e);
-            Err(e)
-        })?;
-
-        // Safety: same conditions than above.
-        unsafe {
-            key_handle.close()?;
-        }
-
-        Ok(psa_import_key::Result {})
+        self.psa_import_key_internal(app_name, op)
     }
 
     fn psa_export_public_key(
@@ -378,54 +193,7 @@ impl Provide for MbedProvider {
         app_name: ApplicationName,
         op: psa_export_public_key::Operation,
     ) -> Result<psa_export_public_key::Result> {
-        info!("Mbed Provider - Export Public Key");
-        let _semaphore_guard = self.key_slot_semaphore.access();
-        let key_name = op.key_name;
-        let key_triple = KeyTriple::new(app_name, ProviderID::MbedCrypto, key_name);
-        let store_handle = self.key_info_store.read().expect("Key store lock poisoned");
-        let key_id = get_key_id(&key_triple, &*store_handle)?;
-
-        let _guard = self
-            .key_handle_mutex
-            .lock()
-            .expect("Grabbing key handle mutex failed");
-
-        let mut key_handle;
-        let mut key_attrs;
-        // Safety:
-        //   * at this point the provider has been instantiated so Mbed Crypto has been initialized
-        //   * self.key_handle_mutex prevents concurrent accesses
-        //   * self.key_slot_semaphore prevents overflowing key slots
-        unsafe {
-            key_handle = KeyHandle::open(key_id)?;
-            key_attrs = key_handle.attributes()?;
-        }
-
-        let buffer_size = utils::psa_export_public_key_size(key_attrs.as_ref())?;
-        let mut buffer = vec![0u8; buffer_size];
-        let mut actual_size = 0;
-
-        let export_status;
-        // Safety: same conditions than above.
-        unsafe {
-            export_status = psa_crypto_binding::psa_export_public_key(
-                key_handle.raw(),
-                buffer.as_mut_ptr(),
-                buffer_size,
-                &mut actual_size,
-            );
-            key_attrs.reset();
-            key_handle.close()?;
-        };
-
-        if export_status != PSA_SUCCESS {
-            error!("Export status: {}", export_status);
-            // Safety: same conditions than above.
-            return Err(utils::convert_status(export_status));
-        }
-
-        buffer.resize(actual_size, 0);
-        Ok(psa_export_public_key::Result { data: buffer })
+        self.psa_export_public_key_internal(app_name, op)
     }
 
     fn psa_destroy_key(
@@ -433,46 +201,7 @@ impl Provide for MbedProvider {
         app_name: ApplicationName,
         op: psa_destroy_key::Operation,
     ) -> Result<psa_destroy_key::Result> {
-        info!("Mbed Provider - Destroy Key");
-        let _semaphore_guard = self.key_slot_semaphore.access();
-        let key_name = op.key_name;
-        let key_triple = KeyTriple::new(app_name, ProviderID::MbedCrypto, key_name);
-        let mut store_handle = self
-            .key_info_store
-            .write()
-            .expect("Key store lock poisoned");
-        let mut local_ids_handle = self.local_ids.write().expect("Local ID lock poisoned");
-        let key_id = get_key_id(&key_triple, &*store_handle)?;
-
-        let _guard = self
-            .key_handle_mutex
-            .lock()
-            .expect("Grabbing key handle mutex failed");
-
-        let key_handle;
-        let destroy_key_status;
-
-        // Safety:
-        //   * at this point the provider has been instantiated so Mbed Crypto has been initialized
-        //   * self.key_handle_mutex prevents concurrent accesses
-        //   * self.key_slot_semaphore prevents overflowing key slots
-        unsafe {
-            key_handle = KeyHandle::open(key_id)?;
-            destroy_key_status = psa_crypto_binding::psa_destroy_key(key_handle.raw());
-        }
-
-        if destroy_key_status == PSA_SUCCESS {
-            remove_key_id(
-                &key_triple,
-                key_id,
-                &mut *store_handle,
-                &mut local_ids_handle,
-            )?;
-            Ok(psa_destroy_key::Result {})
-        } else {
-            error!("Destroy key status: {}", destroy_key_status);
-            Err(utils::convert_status(destroy_key_status))
-        }
+        self.psa_destroy_key_internal(app_name, op)
     }
 
     fn psa_sign_hash(
@@ -480,63 +209,7 @@ impl Provide for MbedProvider {
         app_name: ApplicationName,
         op: psa_sign_hash::Operation,
     ) -> Result<psa_sign_hash::Result> {
-        info!("Mbed Provider - Asym Sign");
-        let _semaphore_guard = self.key_slot_semaphore.access();
-        let key_name = op.key_name;
-        let hash = op.hash;
-        let alg = op.alg;
-        let key_triple = KeyTriple::new(app_name, ProviderID::MbedCrypto, key_name);
-        let store_handle = self.key_info_store.read().expect("Key store lock poisoned");
-        let key_id = get_key_id(&key_triple, &*store_handle)?;
-
-        let _guard = self
-            .key_handle_mutex
-            .lock()
-            .expect("Grabbing key handle mutex failed");
-
-        let mut key_handle;
-        let mut key_attrs;
-        // Safety:
-        //   * at this point the provider has been instantiated so Mbed Crypto has been initialized
-        //   * self.key_handle_mutex prevents concurrent accesses
-        //   * self.key_slot_semaphore prevents overflowing key slots
-        unsafe {
-            key_handle = KeyHandle::open(key_id)?;
-            key_attrs = key_handle.attributes()?;
-        }
-
-        let buffer_size = utils::psa_asymmetric_sign_output_size(key_attrs.as_ref())?;
-        let mut signature = vec![0u8; buffer_size];
-        let mut signature_size: usize = 0;
-
-        let sign_status;
-        // Safety: same conditions than above.
-        unsafe {
-            sign_status = psa_crypto_binding::psa_asymmetric_sign(
-                key_handle.raw(),
-                utils::convert_algorithm(&alg.into())?,
-                hash.as_ptr(),
-                hash.len(),
-                signature.as_mut_ptr(),
-                buffer_size,
-                &mut signature_size,
-            );
-            key_attrs.reset();
-            key_handle.close()?;
-        };
-
-        if sign_status == PSA_SUCCESS {
-            let mut res = psa_sign_hash::Result {
-                signature: Vec::new(),
-            };
-            res.signature.resize(signature_size, 0);
-            res.signature.copy_from_slice(&signature[0..signature_size]);
-
-            Ok(res)
-        } else {
-            error!("Sign status: {}", sign_status);
-            Err(utils::convert_status(sign_status))
-        }
+        self.psa_sign_hash_internal(app_name, op)
     }
 
     fn psa_verify_hash(
@@ -544,48 +217,7 @@ impl Provide for MbedProvider {
         app_name: ApplicationName,
         op: psa_verify_hash::Operation,
     ) -> Result<psa_verify_hash::Result> {
-        info!("Mbed Provider - Asym Verify");
-        let _semaphore_guard = self.key_slot_semaphore.access();
-        let key_name = op.key_name;
-        let hash = op.hash;
-        let alg = op.alg;
-        let signature = op.signature;
-        let key_triple = KeyTriple::new(app_name, ProviderID::MbedCrypto, key_name);
-        let store_handle = self.key_info_store.read().expect("Key store lock poisoned");
-        let key_id = get_key_id(&key_triple, &*store_handle)?;
-
-        let _guard = self
-            .key_handle_mutex
-            .lock()
-            .expect("Grabbing key handle mutex failed");
-
-        let mut key_handle;
-        let mut key_attrs;
-        let verify_status;
-        // Safety:
-        //   * at this point the provider has been instantiated so Mbed Crypto has been initialized
-        //   * self.key_handle_mutex prevents concurrent accesses
-        //   * self.key_slot_semaphore prevents overflowing key slots
-        unsafe {
-            key_handle = KeyHandle::open(key_id)?;
-            key_attrs = key_handle.attributes()?;
-            verify_status = psa_crypto_binding::psa_asymmetric_verify(
-                key_handle.raw(),
-                utils::convert_algorithm(&alg.into())?,
-                hash.as_ptr(),
-                hash.len(),
-                signature.as_ptr(),
-                signature.len(),
-            );
-            key_attrs.reset();
-            key_handle.close()?;
-        }
-
-        if verify_status == PSA_SUCCESS {
-            Ok(psa_verify_hash::Result {})
-        } else {
-            Err(utils::convert_status(verify_status))
-        }
+        self.psa_verify_hash_internal(app_name, op)
     }
 }
